Followup 3: Remove base `DefaultPreferencesStore`. `PublicPreferencesStore` should be moved to inherit from the default store that is being created as part of #157 when that is completed.
